+++ js/src/angular/courseDisplayAndSelect.directive.js
@@ -2,7 +2,7 @@ var BaseCtrl = require('./_base.controller');
 
 function sbCourseDisplayAndSelectDirective() {
   sbCourseDisplayAndSelectCtrl.prototype = Object.create(BaseCtrl.prototype);
-  function sbCourseDisplayAndSelectCtrl($state, $window, scheduleFactory) {
+  function sbCourseDisplayAndSelectCtrl($state, $window, $scope, scheduleFactory) {
     BaseCtrl.call(this, $state, $window);
 
     var vm = this;
@@ -13,11 +13,26 @@ function sbCourseDisplayAndSelectDirective() {
       'GRP': 5,
       'LAB': 7
     };
+    vm.selectAllSections = false;
+    vm.onChangeSelectAllSections = onChangeSelectAllSections;
     vm.setSchedulesStale = setSchedulesStale;
     vm.extractSectionTypeMapping = extractSectionTypeMapping;
 
+    function onChangeSelectAllSections() {
+      if (vm.selectAllSections) {
+        $scope.course.sections.forEach(function(section) {
+          section.selected = true;
+        });
+      } else {
+        $scope.course.sections.forEach(function(section) {
+          section.selected = false;
+        });
+      }
+      setSchedulesStale();
+    }
+
     function setSchedulesStale() {
-      scheduleFactory.setStale();
+      scheduleFactory.setStale(true);
     }
 
     function extractSectionTypeMapping(section) {
@@ -32,6 +47,7 @@ function sbCourseDisplayAndSelectDirective() {
     controller: [
       '$state',
       '$window',
+      '$scope',
       'scheduleFactory',
       sbCourseDisplayAndSelectCtrl
     ],